+++ src/config.js
@@ -1,4 +1,4 @@
-export default {
+export Colors {
   red: 0xf25346,
   white: 0xd8d0d1,
   brown: 0x59332e,
